Commit - 202212032017
(strings.xml) Added the external strings reference file
(Rating.java) Added an enum to handle the ratings class
(MovieSelectionActivity.java) Added the bindings to refer to the inner components of the activity. Also, added the recyclerview set up using a layout manager, adapter, and onItemClickListener. Added the intents to pass the Movie class object to the next screen.
(MovieDetailsActivity.java) Added the bindings to refer to the inner components of the activity. Added the receiving end of the intents that receives the Movie class object. Based on the received data, the data can be filter, sorted, or interpreted.
(Movie.java) Added a Movie class that captures the data needed to represent a movie. This class object is parcelable as it will be passed using intents.
(MainActivity.java) Moved code to more appropriate locations for sections that would no longer fall under this java file.
(Genre.java) Added the genre enum object to differentiate between the different genres using an enum instead of strings.
(dimens.xml) Added the external dimens file to handle the dimensions defined within the app.
(DataGenerator.java) Added the DataGenerator class that is currently used to populate the data needed for the movies used in this example. This section will later need to be usecd to populate the movie data from alternative resources instead of hard coding.
(build.gradle) Updated:
    com.android.application
    com.android.library
    Also, added the android youtube player: implementation 'com.pierfrancescosoffritti.androidyoutubeplayer:core:11.1.0'
    Added, binding references
(AndroidManifest.xml) Added the references needed to display the MovieSelectionActivity and MovieDetailsActivity.
(AdapterList.java) Added the AdapterList class object to be able to utilize the recyclerview in the MovieSelectionActivity.

Added activity_movie_view_v2.xml that will be the recyclerview for an individual movie in the list.

Added activity_movie_selection.xml the shows the list of movies for the user to choose from

Added activity_movie_details.xml that shows details and screen times for a specific movie.

(activity_main_page_v2.xml) Modified the bottom navigation bar to look a lot better.