Commit - 202212041207
This commit is mostly a cleanup...

Removed the following files:
    activity_main_page.xml
    activity_movie_view.xml
    activity_register_old.xml
    activity_seating_chart.xml
    activity_seating_chart_v2.xml
    ic_baseline_add_24.xml
    MovieViewV2Activity.java
    screen_shot_2022_10_28_at_1_48_10_pm.png
    screen_shot_2022_10_28_at_1_54_16_pm.png
    seatingchart.jpg

(MovieDetailsActivity.java) Added the on click listener for the button that was added to the screen. However, it is not fully functioning.
(AndroidManifest.xml) Added the section used for login testing.
(activity_movie_details.xml) Added a button to the bottom of the screen to proceed to picking seats in the movie theater.
(MovieSeatingChartActivity.java) Added the class for activity_seating_chart_v3.xml
(Poster.java) Added a class to get the reference to the movie poster. I have not implement it yet.