Commit - 202212050758
Renamed the movie poster filenames and some icons for more consistency

(MovieDetailsActivity.java) Fixed the chip group and added dynamic on-click listeners for each new chip that's added to the chip-group. Also, added a check to make sure the user has selected a showtime and if not the user is visually notified by a red highlight and cant proceed. Also, added the event class objects instantiation during this process.

(MovieSeatingChartActivity.java) Corrected the activity binding, added the spinner set up, and allocations for the user selecting handicap or reclining seats options.

(build.gradle) Changed the minimum SDK version to 29

(Theater.java) Added the theater class to allow for the creation of specifics regarding the theater. It has been made parcelable so that it can be passed between the activities.

(Seat.java) Added the seat class as a way to define seating options. For example, the user can select handicap, reclining, or normal seats. Also, provides a way to identify the seat number (id) the user has selected. It is also parcelable so the it can be passed around in the activities.

(Event.java) Added the Event java class to coalesce the Movie, Seat, and Theater classes into one single object. It was made parcelable so that it can be passed between activities.

Cleaned | Synced | ReBuilt | Debugged