Commit - 202212050758
(activity_seating_chart_v3.xml) Remade the activity to look a little better and moved the showtimes from the details activity to the seating. That way a user can make selections all on one page.

(MovieDetailsActivity.java) Changed the functionality of this class to allow for the movement of the showtimes to the seating activity. Now the intents only pass the event class with only the movie in it.

(MovieSeatingChartActivity.java) Added an all new activity class to allow for the selection of multiple seats. Also, added the showtimes to this activity.

(Event.java) Added a constructor thats more in line with the changes made. Changes being moving the showtimes from the details page to the seating page.

Cleaned | Synced | ReBuilt | Debugged