Commit - 20221209????
(activity_order_summary.xml) Added this activity to allow for the user to see an itemized list of the running/total cost

(activity_ticket_list.xml) This will be the running list of tickets the usert has purchased and can be retrieved and forwarded as needed.

(AndroidManifest.xml) Added the new activities to the manifest and assign their parent names

(Event.java) Had trouble with the parcelization of the seat list. Had to redo the way I parcelization the data using a writeTypedList instead of writeParcelable.

(MainActivity.java) Added the on-click listeners for the three main buttons on the main page. Also, included the intents to start their own page.

(MovieDetailsActivity.java) Added a temp hardcoded theater for now

(MovieSeatingChartActivity.java) Added the tempChip on click listener to a lambda. Added code comments throughout.

(OrderSummaryActivity.java) Added the java class file for the activity. Also added the buy now on-click listener and the calculate cost function.

(Seat.java) Added the getOptions function to get the seating option in string format for when the user has selected handicap, reclining, and regular seating options.

(TheaterSelectionActivity.java) Added the theater selection activitys java class file. Stock java class file for now.

(ticket_layou.xml) Added the ticket layout activity to display the qr code for the ticket purchased. It has not been fully implemented into the stream yet. However, it does generate a qr code based on passed in information.

(TicketListActivity.java) Added a stock java class file for the ticket list activity. It is not fully functioning at this time as it has to be reworked to add a recyclerview.

(TicketDialogFragment.java) (TicketFragment.java) Both of these have been commented out as they are currently producing errors.

Cleaned | Synced | ReBuilt | Debugged